SystemAggregator: improve upload of diagnostics
We not do not simply pass down the request, but intercept it
and instruct the systems to upload to us. For this a local
HTTP or FTP server must be running and accepting the uploads.

The URL passed to the systems instead is configurable.

After we received all data from the systems, we finally upload
a combined file to the original URL.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>