Improve initial sync behavior between SatelliteController and SatelliteAgent
This now uses a two-step approach which allows improved re-connect detection
and ensures that the dependend objects are not touched while we are not
fully synchronized.

It also introduce timeouts for regular/periodic calls on both sides.

While at, convert the busy-waiting loop in SatelliteAgent to use
signaled condition variables.

As soon as something looks unusual, we try to restart the EVerest stack.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>