SystemAggregator: handle fw update status notification smarter
Just passing the status notifications of the local system is not
sufficient, e.g. when the backend performs actions when specific
events are signaled.

This approach groups different notifications into categories, e.g.
ones that are passed immediately (but only once) or ones which
are not reported before all satellites reported them.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>